api/requirements.txt
numpy==1.11.3
pyyaml==3.12